refactor(ipc): extract main-process IPC handlers from register.ts
