+++ main-src/ipc/register.ts
@@ -44,26 +44,6 @@ setWorkspaceFileIndexReadyBroadcaster((rootNorm) => {
 	}
 });
 
-function createWindowsUnreadOverlayIcon(count: number) {
-	const label = count > 99 ? '99+' : String(count);
-	const fontSize = label.length >= 3 ? 54 : label.length === 2 ? 66 : 78;
-	const svg = `
-<svg xmlns="http://www.w3.org/2000/svg" width="256" height="256" viewBox="0 0 256 256">
-  <circle cx="160" cy="96" r="92" fill="#ef4444"/>
-  <circle cx="160" cy="96" r="82" fill="#ef4444" stroke="white" stroke-width="14"/>
-  <text x="160" y="118" text-anchor="middle" dominant-baseline="middle" font-family="Segoe UI, Arial, sans-serif" font-size="${fontSize}" font-weight="800" fill="white">${label}</text>
-</svg>`;
-	return nativeImage.createFromDataURL(`data:image/svg+xml;charset=utf-8,${encodeURIComponent(svg)}`);
-}
-
-function applyUnreadBadgeToWindow(win: BrowserWindow, count: number): void {
-	const safeCount = Math.max(0, Math.min(999, Math.floor(Number.isFinite(count) ? count : 0)));
-	if (process.platform === 'win32') {
-		win.setOverlayIcon(safeCount > 0 ? createWindowsUnreadOverlayIcon(safeCount) : null, safeCount > 0 ? `${safeCount} unread replies` : '');
-		return;
-	}
-	app.setBadgeCount(safeCount);
-}
 import {
 	getSettings,
 	patchSettings,
@@ -222,6 +202,13 @@ import { registerClipboardHandlers } from './handlers/clipboardHandlers.js';
 import { registerAutoUpdateHandlers } from './handlers/autoUpdateHandlers.js';
 import { registerUsageStatsHandlers } from './handlers/usageStatsHandlers.js';
 import { registerLspHandlers } from './handlers/lspHandlers.js';
+import { registerAppHandlers } from './handlers/appHandlers.js';
+import { registerWorkspaceHandlers } from './handlers/workspaceHandlers.js';
+import { registerFsHandlers } from './handlers/fsHandlers.js';
+import { registerShellHandlers } from './handlers/shellHandlers.js';
+import { registerGitHandlers } from './handlers/gitHandlers.js';
+import { registerBrowserHandlers } from './handlers/browserHandlers.js';
+import { senderWorkspaceRoot } from './agentRuntime.js';
 import { setDelegateContext, clearDelegateContext } from '../agent/toolExecutor.js';
 import {
 	attachManagedAgentEmitter,
@@ -258,35 +245,6 @@ import { generateThreadTitle } from '../threadTitle.js';
 
 const execFileAsync = promisify(execFile);
 
-export function senderWorkspaceRoot(event: { sender: WebContents }): string | null {
-	return getWorkspaceRootForWebContents(event.sender);
-}
-
-function parseAppWindowSurface(raw: unknown): AppWindowSurface | null {
-	return raw === 'agent' || raw === 'editor' ? raw : null;
-}
-
-function broadcastPluginsChanged(): void {
-	for (const win of BrowserWindow.getAllWindows()) {
-		if (win.isDestroyed()) {
-			continue;
-		}
-		try {
-			win.webContents.send('async-shell:pluginsChanged');
-		} catch {
-			/* ignore */
-		}
-	}
-}
-
-function workspaceRootsEqual(a: string | null | undefined, b: string | null | undefined): boolean {
-	if (!a || !b) {
-		return false;
-	}
-	const na = path.resolve(a).replace(/\\/g, '/').toLowerCase();
-	const nb = path.resolve(b).replace(/\\/g, '/').toLowerCase();
-	return na === nb;
-}
 
 /** 可选覆盖工作区根路径（须为已存在目录），否则使用当前窗口绑定的工作区 */
 function resolveWorkspaceScopeForThreads(
@@ -542,188 +500,6 @@ function reverseUnifiedPatch(chunk: string): string | null {
 	}
 }
 
-type ExternalWorkspaceTool = 'vscode' | 'cursor' | 'antigravity' | 'explorer' | 'terminal';
-
-function isExternalWorkspaceTool(value: unknown): value is ExternalWorkspaceTool {
-	return (
-		value === 'vscode' ||
-		value === 'cursor' ||
-		value === 'antigravity' ||
-		value === 'explorer' ||
-		value === 'terminal'
-	);
-}
-
-async function commandOnPath(command: string): Promise<boolean> {
-	try {
-		if (process.platform === 'win32') {
-			await execFileAsync('where.exe', [command], { windowsHide: true });
-		} else {
-			await execFileAsync('which', [command], { windowsHide: true });
-		}
-		return true;
-	} catch {
-		return false;
-	}
-}
-
-function windowsEditorExecutableFallbacks(tool: Extract<ExternalWorkspaceTool, 'vscode' | 'cursor' | 'antigravity'>): string[] {
-	if (process.platform !== 'win32') {
-		return [];
-	}
-	const localAppData = process.env.LOCALAPPDATA;
-	const programFiles = process.env.ProgramFiles;
-	const programFilesX86 = process.env['ProgramFiles(x86)'];
-	switch (tool) {
-		case 'vscode':
-			return [
-				localAppData ? path.join(localAppData, 'Programs', 'Microsoft VS Code', 'Code.exe') : null,
-				programFiles ? path.join(programFiles, 'Microsoft VS Code', 'Code.exe') : null,
-				programFilesX86 ? path.join(programFilesX86, 'Microsoft VS Code', 'Code.exe') : null,
-			].filter((candidate): candidate is string => Boolean(candidate));
-		case 'cursor':
-			return [
-				localAppData ? path.join(localAppData, 'Programs', 'Cursor', 'Cursor.exe') : null,
-				programFiles ? path.join(programFiles, 'Cursor', 'Cursor.exe') : null,
-				programFilesX86 ? path.join(programFilesX86, 'Cursor', 'Cursor.exe') : null,
-			].filter((candidate): candidate is string => Boolean(candidate));
-		case 'antigravity':
-			return [
-				localAppData ? path.join(localAppData, 'Programs', 'Antigravity', 'Antigravity.exe') : null,
-				programFiles ? path.join(programFiles, 'Antigravity', 'Antigravity.exe') : null,
-				programFilesX86 ? path.join(programFilesX86, 'Antigravity', 'Antigravity.exe') : null,
-			].filter((candidate): candidate is string => Boolean(candidate));
-		default:
-			return [];
-	}
-}
-
-type LaunchCommand = {
-	command: string;
-	useShell: boolean;
-};
-
-async function resolveLaunchCommand(candidates: string[]): Promise<LaunchCommand | null> {
-	for (const candidate of candidates) {
-		if (!candidate) {
-			continue;
-		}
-		if (path.isAbsolute(candidate)) {
-			if (fs.existsSync(candidate)) {
-				return { command: candidate, useShell: /\.(cmd|bat)$/i.test(candidate) };
-			}
-			continue;
-		}
-		if (await commandOnPath(candidate)) {
-			return { command: candidate, useShell: process.platform === 'win32' };
-		}
-	}
-	return null;
-}
-
-async function spawnDetachedLaunch(
-	command: string,
-	args: string[],
-	opts?: { cwd?: string; useShell?: boolean; windowsHide?: boolean }
-): Promise<void> {
-	await new Promise<void>((resolve, reject) => {
-		const child = spawn(command, args, {
-			cwd: opts?.cwd,
-			detached: true,
-			stdio: 'ignore',
-			shell: opts?.useShell ?? false,
-			windowsHide: opts?.windowsHide ?? true,
-		});
-		child.once('error', reject);
-		child.once('spawn', () => {
-			child.unref();
-			resolve();
-		});
-	});
-}
-
-async function launchWorkspaceInExternalEditor(
-	tool: Extract<ExternalWorkspaceTool, 'vscode' | 'cursor' | 'antigravity'>,
-	workspaceRoot: string,
-	targetPath?: string,
-	revealLine?: number
-): Promise<boolean> {
-	const commandCandidates = {
-		vscode: ['code'],
-		cursor: ['cursor'],
-		antigravity: ['antigravity'],
-	}[tool];
-	const resolved = await resolveLaunchCommand([
-		...commandCandidates,
-		...windowsEditorExecutableFallbacks(tool),
-	]);
-	if (!resolved) {
-		return false;
-	}
-	const target = targetPath ?? workspaceRoot;
-	const targetArg = targetPath && Number.isFinite(revealLine) && revealLine > 0 ? `${targetPath}:${Math.floor(revealLine)}` : target;
-	await spawnDetachedLaunch(resolved.command, ['-n', targetArg], {
-		cwd: workspaceRoot,
-		useShell: resolved.useShell,
-		windowsHide: true,
-	});
-	return true;
-}
-
-function escapePowerShellLiteral(value: string): string {
-	return value.replace(/'/g, "''");
-}
-
-async function launchWorkspaceInExternalTerminal(workspaceRoot: string): Promise<boolean> {
-	if (process.platform === 'win32') {
-		const wt = await resolveLaunchCommand(['wt']);
-		if (wt) {
-			await spawnDetachedLaunch(wt.command, ['-d', workspaceRoot], {
-				cwd: workspaceRoot,
-				useShell: wt.useShell,
-				windowsHide: true,
-			});
-			return true;
-		}
-		await spawnDetachedLaunch(
-			'powershell.exe',
-			['-NoExit', '-Command', `Set-Location -LiteralPath '${escapePowerShellLiteral(workspaceRoot)}'`],
-			{
-				cwd: workspaceRoot,
-				useShell: false,
-				windowsHide: false,
-			}
-		);
-		return true;
-	}
-	if (process.platform === 'darwin') {
-		await spawnDetachedLaunch('open', ['-a', 'Terminal', workspaceRoot], {
-			cwd: workspaceRoot,
-			useShell: false,
-			windowsHide: true,
-		});
-		return true;
-	}
-	const candidates: Array<{ command: string; args: string[] }> = [
-		{ command: 'x-terminal-emulator', args: ['--working-directory', workspaceRoot] },
-		{ command: 'gnome-terminal', args: [`--working-directory=${workspaceRoot}`] },
-		{ command: 'konsole', args: ['--workdir', workspaceRoot] },
-		{ command: 'xfce4-terminal', args: ['--working-directory', workspaceRoot] },
-	];
-	for (const candidate of candidates) {
-		const resolved = await resolveLaunchCommand([candidate.command]);
-		if (!resolved) {
-			continue;
-		}
-		await spawnDetachedLaunch(resolved.command, candidate.args, {
-			cwd: workspaceRoot,
-			useShell: resolved.useShell,
-			windowsHide: true,
-		});
-		return true;
-	}
-	return false;
-}
 
 function recordTurnTokenUsageStats(
 	modelSelection: string,
@@ -1247,451 +1023,29 @@ export function registerIpc(): void {
 		}
 	});
 
-	ipcMain.handle('async-shell:ping', () => ({ ok: true, message: 'pong' }));
-
-	ipcMain.handle('app:getPaths', () => ({
-		userData: app.getPath('userData'),
-		home: app.getPath('home'),
-	}));
-
-	ipcMain.handle('app:getVersion', () => ({
-		version: app.getVersion(),
-		electron: process.versions.electron ?? '',
-		chrome: process.versions.chrome ?? '',
-		node: process.versions.node ?? '',
-	}));
-
-	ipcMain.handle('app:setUnreadBadgeCount', (event, rawCount: unknown) => {
-		const win = BrowserWindow.fromWebContents(event.sender);
-		if (!win || win.isDestroyed()) {
-			return { ok: false as const };
-		}
-		const count = typeof rawCount === 'number' ? rawCount : Number(rawCount ?? 0);
-		applyUnreadBadgeToWindow(win, count);
-		return { ok: true as const };
-	});
-
-	ipcMain.handle('workspace:pickFolder', async (event) => {
-		const win = BrowserWindow.fromWebContents(event.sender);
-		const r = await dialog.showOpenDialog(win!, {
-			properties: ['openDirectory', 'createDirectory'],
-		});
-		if (r.canceled || !r.filePaths[0]) {
-			return { ok: false as const };
-		}
-		const picked = r.filePaths[0];
-		const resolvedPick = path.resolve(picked);
-		const prev = bindWorkspaceRootToWebContents(event.sender, resolvedPick);
-		if (prev && prev !== resolvedPick) {
-			releaseWorkspaceFileIndexRef(prev);
-		}
-		if (prev !== resolvedPick) {
-			acquireWorkspaceFileIndexRef(resolvedPick);
-		}
-		rememberWorkspace(resolvedPick);
-		return { ok: true as const, path: resolvedPick };
-	});
-
-	ipcMain.handle('workspace:openPath', (event, dirPath: string) => {
-		const t0 = performance.now();
-		try {
-			const resolved = path.resolve(String(dirPath ?? ''));
-			if (!fs.existsSync(resolved)) {
-				return { ok: false as const, error: '路径不存在' };
-			}
-			if (!fs.statSync(resolved).isDirectory()) {
-				return { ok: false as const, error: '不是文件夹' };
-			}
-			const prev = bindWorkspaceRootToWebContents(event.sender, resolved);
-			if (prev && prev !== resolved) {
-				releaseWorkspaceFileIndexRef(prev);
-			}
-			if (prev !== resolved) {
-				acquireWorkspaceFileIndexRef(resolved);
-			}
-			rememberWorkspace(resolved);
-			console.log(`[perf][main] workspace:openPath done in ${(performance.now() - t0).toFixed(1)}ms`);
-			return { ok: true as const, path: resolved };
-		} catch (e) {
-			return { ok: false as const, error: String(e) };
-		}
-	});
-
-	ipcMain.handle('workspace:openInExternalTool', async (event, payload: unknown) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, code: 'no-workspace' as const };
-		}
-		const input = payload as
-			| { tool?: unknown; relPath?: unknown; revealLine?: unknown; revealEndLine?: unknown }
-			| null
-			| undefined;
-		const tool = input?.tool;
-		if (!isExternalWorkspaceTool(tool)) {
-			return { ok: false as const, code: 'unsupported-tool' as const, error: 'unsupported tool' };
-		}
-		const relPath = typeof input?.relPath === 'string' ? input.relPath.trim() : '';
-		let targetPath: string | undefined;
-		if (relPath) {
-			const normalizedRel = relPath.replace(/\\/g, '/');
-			const resolvedTarget = path.resolve(root, normalizedRel);
-			const relativeToRoot = path.relative(root, resolvedTarget);
-			if (relativeToRoot.startsWith('..') || path.isAbsolute(relativeToRoot)) {
-				return { ok: false as const, code: 'outside-workspace' as const, error: 'path is outside workspace' };
-			}
-			targetPath = resolvedTarget;
-		}
-		const revealLine = typeof input?.revealLine === 'number' ? input.revealLine : undefined;
-		try {
-			if (tool === 'explorer') {
-				const err = await shell.openPath(targetPath ?? root);
-				return err
-					? ({ ok: false as const, code: 'launch-failed' as const, error: err } as const)
-					: ({ ok: true as const } as const);
-			}
-			if (tool === 'terminal') {
-				const ok = await launchWorkspaceInExternalTerminal(root);
-				return ok
-					? ({ ok: true as const } as const)
-					: ({ ok: false as const, code: 'tool-unavailable' as const } as const);
-			}
-			const ok = await launchWorkspaceInExternalEditor(tool, root, targetPath, revealLine);
-			return ok
-				? ({ ok: true as const } as const)
-				: ({ ok: false as const, code: 'tool-unavailable' as const } as const);
-		} catch (e) {
-			return {
-				ok: false as const,
-				code: 'launch-failed' as const,
-				error: e instanceof Error ? e.message : String(e),
-			};
-		}
-	});
-
-	ipcMain.handle('workspace:listRecents', () => {
-		const t0 = performance.now();
-		const seen = new Set<string>();
-		const candidates = [
-			...getRecentWorkspaces(),
-			...listThreadWorkspaceRoots({ onlyWithUserMessages: true }),
-		];
-		const paths = candidates.filter((p) => {
-			const key = path.resolve(p).replace(/\\/g, '/').toLowerCase();
-			if (seen.has(key)) {
-				return false;
-			}
-			seen.add(key);
-			try {
-				return fs.existsSync(p) && fs.statSync(p).isDirectory();
-			} catch {
-				return false;
-			}
-		});
-		console.log(`[perf][main] workspace:listRecents done in ${(performance.now() - t0).toFixed(1)}ms, count=${paths.length}`);
-		return { paths };
-	});
-
-	ipcMain.handle('workspace:removeRecent', (_e, dirPath: string) => {
-		try {
-			const resolved = path.resolve(String(dirPath ?? ''));
-			removeRecentWorkspace(resolved);
-			return { ok: true as const };
-		} catch (e) {
-			return { ok: false as const, error: e instanceof Error ? e.message : String(e) };
-		}
-	});
-
-	ipcMain.handle('workspace:get', (event) => ({ root: senderWorkspaceRoot(event) }));
-
-	ipcMain.handle('workspace:searchSymbols', async (event, query: string) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: true as const, hits: [] as { name: string; path: string; line: number; kind: string }[] };
-		}
-		const rootNorm = path.resolve(root);
-		await ensureSymbolIndexLoaded(rootNorm);
-		const hits = searchWorkspaceSymbols(String(query ?? ''), 80, rootNorm);
-		return { ok: true as const, hits };
-	});
-
+	registerAppHandlers();
+	registerWorkspaceHandlers();
+	registerFsHandlers();
+	registerShellHandlers();
+	registerGitHandlers();
+	registerBrowserHandlers();
 	registerLspHandlers();
+	registerClipboardHandlers();
+	registerUsageStatsHandlers();
+	registerAutoUpdateHandlers();
 
-	ipcMain.handle('workspace:closeFolder', async (event) => {
-		const root = senderWorkspaceRoot(event);
-		bindWorkspaceRootToWebContents(event.sender, null);
-		if (root) {
-			releaseWorkspaceFileIndexRef(root);
-			clearGitContextCacheForRoot(root);
-		}
-		await disposeTsLspSessionForWebContents(event.sender);
-		return { ok: true as const };
-	});
 
-	ipcMain.handle('app:newWindow', () => {
-		createAppWindow({ blank: true, surface: 'agent' });
-		return { ok: true as const };
-	});
 
-	ipcMain.handle('app:newEditorWindow', () => {
-		createAppWindow({ blank: true, surface: 'editor' });
-		return { ok: true as const };
-	});
 
-	ipcMain.handle('app:windowSurfaceStatus', (event, rawSurface: unknown) => {
-		const surface = parseAppWindowSurface(rawSurface);
-		if (!surface) {
-			return { ok: false as const, error: 'invalid-surface' as const };
-		}
-		const existing = findAppWindowBySurface(surface, {
-			workspaceRoot: senderWorkspaceRoot(event),
-			excludeWebContentsId: event.sender.id,
-		});
-		return { ok: true as const, exists: !!existing };
-	});
 
-	ipcMain.handle('app:openOrFocusWindowSurface', (event, rawSurface: unknown) => {
-		const surface = parseAppWindowSurface(rawSurface);
-		if (!surface) {
-			return { ok: false as const, error: 'invalid-surface' as const };
-		}
-		const initialWorkspace = senderWorkspaceRoot(event);
-		const existing = findAppWindowBySurface(surface, {
-			workspaceRoot: initialWorkspace,
-			excludeWebContentsId: event.sender.id,
-		});
-		if (existing) {
-			focusAppWindow(existing);
-			return { ok: true as const, action: 'focused' as const };
-		}
-		createAppWindow({
-			surface,
-			initialWorkspace,
-		});
-		return { ok: true as const, action: 'created' as const };
-	});
 
-	ipcMain.handle('app:windowGetState', (event) => {
-		const win = BrowserWindow.fromWebContents(event.sender);
-		if (!win) {
-			return { ok: false as const, error: 'no-window' as const };
-		}
-		return { ok: true as const, maximized: win.isMaximized() };
-	});
 
-	ipcMain.handle('app:windowMinimize', (event) => {
-		const win = BrowserWindow.fromWebContents(event.sender);
-		if (!win) {
-			return { ok: false as const, error: 'no-window' as const };
-		}
-		win.minimize();
-		return { ok: true as const };
-	});
 
-	ipcMain.handle('app:windowToggleMaximize', (event) => {
-		const win = BrowserWindow.fromWebContents(event.sender);
-		if (!win) {
-			return { ok: false as const, error: 'no-window' as const };
-		}
-		if (win.isMaximized()) {
-			win.unmaximize();
-		} else {
-			win.maximize();
-		}
-		return { ok: true as const };
-	});
 
-	ipcMain.handle('app:windowClose', (event) => {
-		const win = BrowserWindow.fromWebContents(event.sender);
-		if (!win) {
-			return { ok: false as const, error: 'no-window' as const };
-		}
-		win.close();
-		return { ok: true as const };
-	});
 
-	ipcMain.handle('app:quit', () => {
-		app.quit();
-		return { ok: true as const };
-	});
 
-	ipcMain.handle('fs:pickOpenFile', async (event) => {
-		const win = BrowserWindow.fromWebContents(event.sender);
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'no-workspace' as const };
-		}
-		const r = await dialog.showOpenDialog(win ?? undefined, {
-			properties: ['openFile'],
-			defaultPath: root,
-		});
-		if (r.canceled || !r.filePaths[0]) {
-			return { ok: false as const, canceled: true as const };
-		}
-		const picked = path.resolve(r.filePaths[0]);
-		if (!isPathInsideRoot(picked, root)) {
-			return { ok: false as const, error: 'outside-workspace' as const };
-		}
-		const rel = path.relative(root, picked).split(path.sep).join('/');
-		return { ok: true as const, relPath: rel };
-	});
 
-	ipcMain.handle(
-		'fs:pickSaveFile',
-		async (event, opts?: { defaultName?: string; title?: string }) => {
-			const win = BrowserWindow.fromWebContents(event.sender);
-			const root = senderWorkspaceRoot(event);
-			if (!root) {
-				return { ok: false as const, error: 'no-workspace' as const };
-			}
-			const defaultName = typeof opts?.defaultName === 'string' ? opts.defaultName : 'Untitled.txt';
-			const r = await dialog.showSaveDialog(win ?? undefined, {
-				title: typeof opts?.title === 'string' ? opts.title : 'Save',
-				defaultPath: path.join(root, path.basename(defaultName)),
-			});
-			if (r.canceled || !r.filePath) {
-				return { ok: false as const, canceled: true as const };
-			}
-			const picked = path.resolve(r.filePath);
-			if (!isPathInsideRoot(picked, root)) {
-				return { ok: false as const, error: 'outside-workspace' as const };
-			}
-			const rel = path.relative(root, picked).split(path.sep).join('/');
-			return { ok: true as const, relPath: rel };
-		}
-	);
 
-	ipcMain.handle('workspace:listFiles', async (event) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'no-workspace' as const };
-		}
-		try {
-			const paths = await ensureWorkspaceFileIndex(root);
-			return { ok: true as const, paths };
-		} catch {
-			return { ok: false as const, error: 'read-failed' as const };
-		}
-	});
-
-	ipcMain.handle(
-		'workspace:searchFiles',
-		async (event, opts: { query?: string; gitChangedPaths?: string[]; limit?: number } | undefined) => {
-			const root = senderWorkspaceRoot(event);
-			if (!root) {
-				return { ok: false as const, items: [] };
-			}
-			try {
-				const items = await searchWorkspaceFiles(
-					root,
-					opts?.query ?? '',
-					opts?.gitChangedPaths ?? [],
-					opts?.limit ?? 60
-				);
-				return { ok: true as const, items };
-			} catch {
-				return { ok: false as const, items: [] };
-			}
-		}
-	);
-
-	ipcMain.handle('browser:getConfig', async (event) => {
-		const hostId = resolveBrowserHostIdForSenderId(event.sender.id);
-		const partition = browserPartitionForHostId(hostId);
-		const payload = await getBrowserSidebarConfigPayloadForHostId(hostId);
-		return {
-			ok: true as const,
-			partition,
-			config: payload.config,
-			defaultUserAgent: payload.defaultUserAgent,
-		};
-	});
-
-	ipcMain.handle('browser:setConfig', async (event, rawConfig: unknown) => {
-		const hostId = resolveBrowserHostIdForSenderId(event.sender.id);
-		const result = await setBrowserSidebarConfigForHostId(hostId, rawConfig);
-		if (result.ok) {
-			sendApplyConfigToDetachedBrowserWindowIfOpen(hostId, result.config, result.defaultUserAgent);
-		}
-		return result;
-	});
-
-	const SETTINGS_OPEN_NAV_IDS = new Set([
-		'general',
-		'appearance',
-		'editor',
-		'plan',
-		'team',
-		'bots',
-		'agents',
-		'models',
-		'plugins',
-		'rules',
-		'tools',
-		'indexing',
-		'autoUpdate',
-		'browser',
-	]);
-
-	ipcMain.handle('app:requestOpenSettings', async (event, payload: unknown) => {
-		const navRaw =
-			payload && typeof payload === 'object' && typeof (payload as { nav?: unknown }).nav === 'string'
-				? String((payload as { nav: string }).nav).trim()
-				: '';
-		const nav = navRaw || 'general';
-		if (!SETTINGS_OPEN_NAV_IDS.has(nav)) {
-			return { ok: false as const, error: 'invalid-nav' as const };
-		}
-		const hostId = resolveBrowserHostIdForSenderId(event.sender.id);
-		const mainContents = webContents.fromId(hostId);
-		if (!mainContents || mainContents.isDestroyed()) {
-			return { ok: false as const, error: 'no-host' as const };
-		}
-		mainContents.send('async-shell:openSettingsNav', nav);
-		const win = BrowserWindow.fromWebContents(mainContents);
-		if (win && !win.isDestroyed()) {
-			if (win.isMinimized()) {
-				win.restore();
-			}
-			win.show();
-			win.focus();
-		}
-		return { ok: true as const };
-	});
-
-	ipcMain.handle('browser:syncState', async (event, rawState: unknown) => {
-		const hostId = resolveBrowserHostIdForSenderId(event.sender.id);
-		syncBrowserCaptureBindingsForHostId(hostId, rawState);
-		const state = updateBrowserRuntimeStateForHostId(hostId, rawState);
-		return {
-			ok: true as const,
-			state,
-		};
-	});
-
-	ipcMain.handle('browser:getState', async (event) => {
-		const hostId = resolveBrowserHostIdForSenderId(event.sender.id);
-		return {
-			ok: true as const,
-			state: getBrowserRuntimeStateForHostId(hostId),
-		};
-	});
-
-	ipcMain.handle('browser:commandResult', async (event, payload: unknown) => {
-		const hostId = resolveBrowserHostIdForSenderId(event.sender.id);
-		return {
-			ok: resolveBrowserCommandResultForHostId(hostId, payload),
-		};
-	});
-
-	ipcMain.handle('browser:windowReady', async (event) => {
-		markBrowserWindowReadyForSenderId(event.sender.id);
-		return { ok: true as const };
-	});
-
-	ipcMain.handle('browser:openWindow', async (event) => {
-		const hostId = resolveBrowserHostIdForSenderId(event.sender.id);
-		return { ok: await openBrowserWindowForHostId(hostId) };
-	});
 
 	const COMPOSER_ATTACH_MAX_BYTES = 8 * 1024 * 1024;
 
@@ -2586,7 +1940,7 @@ export function registerIpc(): void {
 		return { ok: true as const, fileStates: t.fileStates ?? {} };
 	});
 
-	registerUsageStatsHandlers();
+
 
 	ipcMain.handle('threads:messages', (_e, threadId: string) => {
 		const t = getThread(threadId);
@@ -3321,25 +2675,6 @@ export function registerIpc(): void {
 		}
 	);
 
-	ipcMain.handle('fs:readFile', (event, relPath: string) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'no-workspace' as const };
-		}
-		const full = resolveWorkspacePath(relPath, root);
-		return { ok: true as const, content: fs.readFileSync(full, 'utf8') };
-	});
-
-	ipcMain.handle('fs:writeFile', (event, relPath: string, content: string) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'no-workspace' as const };
-		}
-		const full = resolveWorkspacePath(relPath, root);
-		fs.mkdirSync(path.dirname(full), { recursive: true });
-		fs.writeFileSync(full, content, 'utf8');
-		return { ok: true as const };
-	});
 
 	ipcMain.handle('agent:keepLastTurn', (_e, threadId: string) => {
 		agentRevertSnapshotsByThread.delete(threadId);
@@ -3531,512 +2866,11 @@ ipcMain.handle(
 		return { ok: true as const, reverted: true };
 	});
 
-	ipcMain.handle('fs:listDir', (event, relPath: string) => {
-		try {
-			const root = senderWorkspaceRoot(event);
-			if (!root) {
-				return { ok: false as const, error: 'No workspace' };
-			}
-			const normalized = typeof relPath === 'string' ? relPath.trim() : '';
-			const full = normalized ? resolveWorkspacePath(normalized, root) : root;
-			if (!isPathInsideRoot(full, root) && full !== root) {
-				return { ok: false as const, error: 'Bad path' };
-			}
-			const entries = fs.readdirSync(full, { withFileTypes: true });
-			const list = entries
-				.map((ent) => {
-					const joined = normalized ? path.join(normalized, ent.name) : ent.name;
-					const relSlash = joined.split(path.sep).join('/');
-					return { name: ent.name, isDirectory: ent.isDirectory(), rel: relSlash };
-				})
-				.sort((a, b) => {
-					if (a.isDirectory !== b.isDirectory) {
-						return a.isDirectory ? -1 : 1;
-					}
-					return a.name.localeCompare(b.name, undefined, { sensitivity: 'base' });
-				});
-			return { ok: true as const, entries: list };
-		} catch (e) {
-			return { ok: false as const, error: String(e) };
-		}
-	});
 
-	ipcMain.handle('shell:revealInFolder', (event, relPath: string) => {
-		try {
-			const root = senderWorkspaceRoot(event);
-			if (!root) {
-				return { ok: false as const, error: 'No workspace' };
-			}
-			const rel = String(relPath ?? '').trim();
-			if (!rel) {
-				return { ok: false as const, error: 'empty path' };
-			}
-			const full = resolveWorkspacePath(rel, root);
-			if (!fs.existsSync(full)) {
-				return { ok: false as const, error: 'not found' };
-			}
-			const st = fs.statSync(full);
-			if (st.isDirectory()) {
-				void shell.openPath(full);
-			} else {
-				shell.showItemInFolder(full);
-			}
-			return { ok: true as const };
-		} catch (e) {
-			return { ok: false as const, error: e instanceof Error ? e.message : String(e) };
-		}
-	});
 
-	ipcMain.handle('shell:revealAbsolutePath', async (_e, rawPath: string) => {
-		try {
-			const target = String(rawPath ?? '').trim();
-			if (!target) {
-				return { ok: false as const, error: 'empty path' };
-			}
-			const full = path.resolve(target);
-			if (!fs.existsSync(full)) {
-				return { ok: false as const, error: 'not found' };
-			}
-			const st = fs.statSync(full);
-			if (process.platform === 'win32') {
-				try {
-					const args = st.isDirectory() ? [full] : [`/select,${full}`];
-					const child = spawn('explorer.exe', args, {
-						detached: true,
-						stdio: 'ignore',
-						windowsHide: false,
-					});
-					child.unref();
-					return { ok: true as const };
-				} catch {
-					/* fall through */
-				}
-			}
-			if (process.platform === 'darwin' && !st.isDirectory()) {
-				try {
-					await execFileAsync('open', ['-R', full], { windowsHide: true });
-					return { ok: true as const };
-				} catch {
-					/* fall through */
-				}
-			}
-			if (st.isDirectory()) {
-				const err = await shell.openPath(full);
-				return err ? ({ ok: false as const, error: err } as const) : ({ ok: true as const } as const);
-			}
-			shell.showItemInFolder(full);
-			return { ok: true as const };
-		} catch (e) {
-			return { ok: false as const, error: e instanceof Error ? e.message : String(e) };
-		}
-	});
 
-	ipcMain.handle('shell:openDefault', async (event, relPath: string) => {
-		try {
-			const root = senderWorkspaceRoot(event);
-			const rel = String(relPath ?? '').trim();
-			if (!rel) {
-				return { ok: false as const, error: 'empty path' };
-			}
-			let full = rel;
-			if (!path.isAbsolute(full)) {
-				if (!root) {
-					return { ok: false as const, error: 'No workspace' };
-				}
-				full = resolveWorkspacePath(rel, root);
-			}
-			if (!fs.existsSync(full) || !fs.statSync(full).isFile()) {
-				return { ok: false as const, error: 'not a file' };
-			}
-			const err = await shell.openPath(full);
-			return err ? ({ ok: false as const, error: err } as const) : ({ ok: true as const } as const);
-		} catch (e) {
-			return { ok: false as const, error: e instanceof Error ? e.message : String(e) };
-		}
-	});
 
-	ipcMain.handle('shell:openInBrowser', async (event, relPath: string) => {
-		try {
-			const root = senderWorkspaceRoot(event);
-			if (!root) {
-				return { ok: false as const, error: 'No workspace' };
-			}
-			const rel = String(relPath ?? '').trim();
-			if (!rel) {
-				return { ok: false as const, error: 'empty path' };
-			}
-			const full = resolveWorkspacePath(rel, root);
-			if (!fs.existsSync(full) || !fs.statSync(full).isFile()) {
-				return { ok: false as const, error: 'not a file' };
-			}
-			const ext = path.extname(full).toLowerCase();
-			if (!['.html', '.htm', '.svg'].includes(ext)) {
-				return { ok: false as const, error: 'unsupported type' };
-			}
-			await shell.openExternal(pathToFileURL(full).href);
-			return { ok: true as const };
-		} catch (e) {
-			return { ok: false as const, error: e instanceof Error ? e.message : String(e) };
-		}
-	});
 
-	ipcMain.handle('shell:openExternalUrl', async (_event, url: string) => {
-		try {
-			const trimmed = String(url ?? '').trim();
-			if (!trimmed) {
-				return { ok: false as const, error: 'empty url' };
-			}
-			let parsed: URL;
-			try {
-				parsed = new URL(trimmed);
-			} catch {
-				return { ok: false as const, error: 'invalid url' };
-			}
-			if (parsed.protocol !== 'http:' && parsed.protocol !== 'https:' && parsed.protocol !== 'mailto:') {
-				return { ok: false as const, error: 'unsupported protocol' };
-			}
-			await shell.openExternal(parsed.toString());
-			return { ok: true as const };
-		} catch (e) {
-			return { ok: false as const, error: e instanceof Error ? e.message : String(e) };
-		}
-	});
-
-	registerClipboardHandlers();
-
-	ipcMain.handle('fs:renameEntry', (event, relPath: string, newName: string) => {
-		try {
-			const root = senderWorkspaceRoot(event);
-			if (!root) {
-				return { ok: false as const, error: 'No workspace' };
-			}
-			const fromRel = String(relPath ?? '').trim();
-			if (!fromRel) {
-				return { ok: false as const, error: 'empty path' };
-			}
-			const fromFull = resolveWorkspacePath(fromRel, root);
-			if (!fs.existsSync(fromFull)) {
-				return { ok: false as const, error: 'not found' };
-			}
-			const base = path.basename(String(newName ?? '').trim());
-			if (!base || base === '.' || base === '..' || base.includes('/') || base.includes('\\')) {
-				return { ok: false as const, error: 'bad name' };
-			}
-			const toFull = path.join(path.dirname(fromFull), base);
-			if (!isPathInsideRoot(toFull, root)) {
-				return { ok: false as const, error: 'escapes workspace' };
-			}
-			if (fs.existsSync(toFull)) {
-				return { ok: false as const, error: 'destination exists' };
-			}
-			fs.renameSync(fromFull, toFull);
-			const newRel = path.relative(root, toFull).split(path.sep).join('/');
-			return { ok: true as const, newRel };
-		} catch (e) {
-			return { ok: false as const, error: e instanceof Error ? e.message : String(e) };
-		}
-	});
-
-	ipcMain.handle('fs:removeEntry', (event, relPath: string, recursive?: unknown) => {
-		try {
-			const root = senderWorkspaceRoot(event);
-			if (!root) {
-				return { ok: false as const, error: 'No workspace' };
-			}
-			const rel = String(relPath ?? '').trim();
-			if (!rel) {
-				return { ok: false as const, error: 'empty path' };
-			}
-			const full = resolveWorkspacePath(rel, root);
-			if (!fs.existsSync(full)) {
-				return { ok: false as const, error: 'not found' };
-			}
-			const st = fs.statSync(full);
-			if (st.isDirectory()) {
-				if (recursive === true) {
-					fs.rmSync(full, { recursive: true, force: true });
-				} else {
-					fs.rmdirSync(full);
-				}
-			} else {
-				fs.unlinkSync(full);
-			}
-			return { ok: true as const };
-		} catch (e) {
-			return { ok: false as const, error: e instanceof Error ? e.message : String(e) };
-		}
-	});
-
-	ipcMain.handle('git:status', async (event) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'No workspace' };
-		}
-		return await gitService.withGitWorkspaceRootAsync(root, async () => {
-			try {
-				const probe = await gitService.gitProbeContext();
-				if (!probe.ok) {
-					return { ok: false as const, error: probe.message };
-				}
-				const gitTop = probe.topLevel;
-				const [porcelain, branch] = await Promise.all([
-					gitService.gitStatusPorcelain(),
-					gitService.gitBranch(),
-				]);
-				const lines = porcelain ? porcelain.split('\n').filter(Boolean) : [];
-				const rawPathStatus = gitService.parseGitPathStatus(lines);
-				const rawOrdered = gitService.listPorcelainPaths(lines);
-				const pathStatus: Record<string, gitService.PathStatusEntry> = {};
-				for (const [repoRel, entry] of Object.entries(rawPathStatus)) {
-					const wsRel = gitService.workspaceRelativeFromRepoRelative(repoRel, root, gitTop);
-					if (wsRel) {
-						pathStatus[wsRel] = entry;
-					}
-				}
-				const changedPaths: string[] = [];
-				const seen = new Set<string>();
-				for (const repoRel of rawOrdered) {
-					const wsRel = gitService.workspaceRelativeFromRepoRelative(repoRel, root, gitTop);
-					if (wsRel && !seen.has(wsRel)) {
-						seen.add(wsRel);
-						changedPaths.push(wsRel);
-					}
-				}
-				return { ok: true as const, branch, lines, pathStatus, changedPaths };
-			} catch (e) {
-				return {
-					ok: false as const,
-					error: gitService.normalizeGitFailureMessage(e, 'Failed to load changes'),
-				};
-			}
-		});
-	});
-
-	ipcMain.handle('git:fullStatus', async (event) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'No workspace' };
-		}
-		return await gitService.withGitWorkspaceRootAsync(root, async () => {
-			try {
-				const probe = await gitService.gitProbeContext();
-				if (!probe.ok) {
-					return { ok: false as const, error: probe.message };
-				}
-				const gitTop = probe.topLevel;
-				const [porcelain, branchListPack] = await Promise.all([
-					gitService.gitStatusPorcelain(),
-					gitService.gitListLocalBranches(),
-				]);
-				const lines = porcelain ? porcelain.split('\n').filter(Boolean) : [];
-				const rawPathStatus = gitService.parseGitPathStatus(lines);
-				const rawOrdered = gitService.listPorcelainPaths(lines);
-				const pathStatus: Record<string, gitService.PathStatusEntry> = {};
-				for (const [repoRel, entry] of Object.entries(rawPathStatus)) {
-					const wsRel = gitService.workspaceRelativeFromRepoRelative(repoRel, root, gitTop);
-					if (wsRel) {
-						pathStatus[wsRel] = entry;
-					}
-				}
-				const changedPaths: string[] = [];
-				const seen = new Set<string>();
-				for (const repoRel of rawOrdered) {
-					const wsRel = gitService.workspaceRelativeFromRepoRelative(repoRel, root, gitTop);
-					if (wsRel && !seen.has(wsRel)) {
-						seen.add(wsRel);
-						changedPaths.push(wsRel);
-					}
-				}
-				const branch = branchListPack.current?.trim() ? branchListPack.current : 'master';
-				const branches = branchListPack.branches;
-				const current = branchListPack.current;
-				let previews: Record<string, gitService.DiffPreview> = {};
-				if (changedPaths.length > 0) {
-					const fullDiffRaw = await gitService.gitDiffHeadUnified(root);
-					previews = await gitService.buildDiffPreviewsMap(
-						changedPaths,
-						fullDiffRaw,
-						root,
-						gitTop,
-						{ maxChars: 4_000 }
-					);
-				}
-				return {
-					ok: true as const,
-					branch,
-					lines,
-					pathStatus,
-					changedPaths,
-					branches,
-					current,
-					previews,
-				};
-			} catch (e) {
-				return {
-					ok: false as const,
-					error: gitService.normalizeGitFailureMessage(e, 'Failed to load changes'),
-				};
-			}
-		});
-	});
-
-	ipcMain.handle('git:stageAll', async (event) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'No workspace' };
-		}
-		try {
-			return await gitService.withGitWorkspaceRootAsync(root, async () => {
-				await gitService.gitStageAll();
-				return { ok: true as const };
-			});
-		} catch (e) {
-			return { ok: false as const, error: String(e) };
-		}
-	});
-
-	ipcMain.handle('git:commit', async (event, message: string) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'No workspace' };
-		}
-		try {
-			return await gitService.withGitWorkspaceRootAsync(root, async () => {
-				await gitService.gitCommit(message);
-				return { ok: true as const };
-			});
-		} catch (e) {
-			return { ok: false as const, error: String(e) };
-		}
-	});
-
-	ipcMain.handle('git:push', async (event) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'No workspace' };
-		}
-		try {
-			return await gitService.withGitWorkspaceRootAsync(root, async () => {
-				await gitService.gitPush();
-				return { ok: true as const };
-			});
-		} catch (e) {
-			return { ok: false as const, error: String(e) };
-		}
-	});
-
-	ipcMain.handle('git:diffPreviews', async (event, relPaths: string[]) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'No workspace' };
-		}
-		const list = Array.isArray(relPaths) ? relPaths : [];
-		try {
-			return await gitService.withGitWorkspaceRootAsync(root, async () => {
-				const probe = await gitService.gitProbeContext();
-				if (!probe.ok) {
-					return { ok: false as const, error: probe.message };
-				}
-				const fullDiffRaw = await gitService.gitDiffHeadUnified(root);
-				const previews = await gitService.buildDiffPreviewsMap(list, fullDiffRaw, root, probe.topLevel, { maxChars: 4_000 });
-				return { ok: true as const, previews };
-			});
-		} catch (e) {
-			return { ok: false as const, error: String(e) };
-		}
-	});
-
-	ipcMain.handle(
-		'git:diffPreview',
-		async (event, payload: { relPath?: string; full?: boolean; maxChars?: number | null }) => {
-			const root = senderWorkspaceRoot(event);
-			if (!root) {
-				return { ok: false as const, error: 'No workspace' };
-			}
-			const relPath = String(payload?.relPath ?? '').trim();
-			if (!relPath) {
-				return { ok: false as const, error: 'Bad path' };
-			}
-			try {
-				const preview = await gitService.getDiffPreview(
-					relPath,
-					{
-						maxChars: payload?.full ? null : payload?.maxChars,
-					},
-					root
-				);
-				return { ok: true as const, preview };
-			} catch (e) {
-				return { ok: false as const, error: String(e) };
-			}
-		}
-	);
-
-	ipcMain.handle('git:listBranches', async (event) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'No workspace' };
-		}
-		return await gitService.withGitWorkspaceRootAsync(root, async () => {
-			try {
-				const probe = await gitService.gitProbeContext();
-				if (!probe.ok) {
-					return { ok: false as const, error: probe.message };
-				}
-				const { branches, current } = await gitService.gitListLocalBranches();
-				return { ok: true as const, branches, current };
-			} catch (e) {
-				return {
-					ok: false as const,
-					error: gitService.normalizeGitFailureMessage(e, 'Could not load branches'),
-				};
-			}
-		});
-	});
-
-	ipcMain.handle('git:checkoutBranch', async (event, branch: string) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'No workspace' };
-		}
-		return gitService.withGitWorkspaceRootAsync(root, async () => {
-			try {
-				const probe = await gitService.gitProbeContext();
-				if (!probe.ok) {
-					return { ok: false as const, error: probe.message };
-				}
-				await gitService.gitSwitchBranch(typeof branch === 'string' ? branch : '');
-				return { ok: true as const };
-			} catch (e) {
-				return {
-					ok: false as const,
-					error: gitService.normalizeGitFailureMessage(e, 'Could not switch branch'),
-				};
-			}
-		});
-	});
-
-	ipcMain.handle('git:createBranch', async (event, name: string) => {
-		const root = senderWorkspaceRoot(event);
-		if (!root) {
-			return { ok: false as const, error: 'No workspace' };
-		}
-		return gitService.withGitWorkspaceRootAsync(root, async () => {
-			try {
-				const probe = await gitService.gitProbeContext();
-				if (!probe.ok) {
-					return { ok: false as const, error: probe.message };
-				}
-				await gitService.gitCreateBranchAndSwitch(typeof name === 'string' ? name : '');
-				return { ok: true as const };
-			} catch (e) {
-				return {
-					ok: false as const,
-					error: gitService.normalizeGitFailureMessage(e, 'Could not create branch'),
-				};
-			}
-		});
-	});
 
 	ipcMain.handle(
 		'plan:save',
@@ -4232,5 +3066,5 @@ ipcMain.handle(
 	});
 
 	/** 自动更新：检查更新 */
-	registerAutoUpdateHandlers();
+
 }
